+++ packages/sdks-tests/src/specs/slot-with-symbol.ts
@@ -1,0 +1,342 @@
+export const SLOT_WITH_SYMBOL = {
+  ownerId: 'ad30f9a246614faaa6a03374f83554c9',
+  lastUpdateBy: null,
+  createdDate: 1705851595269,
+  id: 'c679b54217cc462484fb3d6c5645de4f',
+  '@version': 4,
+  name: 'make-a-slot',
+  modelId: '17c6065109ef4062ba083f5741f4ee6a',
+  published: 'published',
+  meta: {
+    hasLinks: false,
+    kind: 'page',
+    lastPreviewUrl:
+      'http://localhost:5173/make-a-slot?builder.space=ad30f9a246614faaa6a03374f83554c9&builder.cachebust=true&builder.preview=page&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.page=c679b54217cc462484fb3d6c5645de4f&builder.overrides.c679b54217cc462484fb3d6c5645de4f=c679b54217cc462484fb3d6c5645de4f&builder.overrides.page:/make-a-slot=c679b54217cc462484fb3d6c5645de4f',
+    symbolsUsed: {
+      '6926ca2352d24d3286f1af71561cbb01': true,
+    },
+  },
+  priority: -93,
+  query: [
+    {
+      '@type': '@builder.io/core:Query',
+      property: 'urlPath',
+      operator: 'is',
+      value: '/make-a-slot',
+    },
+  ],
+  data: {
+    inputs: [],
+    title: 'make-a-slot',
+    themeId: false,
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-f9f470d1e090475a9861cc3d72cd1570',
+        component: {
+          name: 'Symbol',
+          options: {
+            symbol: {
+              data: {
+                children: [
+                  {
+                    '@type': '@builder.io/sdk:Element',
+                    '@version': 2,
+                    id: 'builder-714cee55b74742ee8b999d70b5bbd9f9',
+                    component: {
+                      name: 'Symbol',
+                      options: {
+                        symbol: {
+                          data: {
+                            children: [
+                              {
+                                '@type': '@builder.io/sdk:Element',
+                                '@version': 2,
+                                id: 'builder-767c90ca7dd34fab889e5d4de9be0a0e',
+                                component: {
+                                  name: 'Text',
+                                  options: {
+                                    text: 'This is called recursion!',
+                                  },
+                                },
+                                responsiveStyles: {
+                                  large: {
+                                    display: 'flex',
+                                    flexDirection: 'column',
+                                    position: 'relative',
+                                    flexShrink: '0',
+                                    boxSizing: 'border-box',
+                                    marginTop: '20px',
+                                    lineHeight: 'normal',
+                                    height: 'auto',
+                                  },
+                                },
+                              },
+                            ],
+                          },
+                          model: 'symbol',
+                          entry: '6926ca2352d24d3286f1af71561cbb01',
+                          content: {
+                            ownerId: 'ad30f9a246614faaa6a03374f83554c9',
+                            lastUpdateBy: null,
+                            createdDate: 1705902347406,
+                            id: '6926ca2352d24d3286f1af71561cbb01',
+                            '@version': 4,
+                            name: 'hello world',
+                            modelId: '5e1209efea0045f58d9b871d81b652fa',
+                            published: 'published',
+                            meta: {
+                              lastPreviewUrl:
+                                'https://preview.builder.codes?model=symbol&previewing=true&apiKey=ad30f9a246614faaa6a03374f83554c9&builder.space=ad30f9a246614faaa6a03374f83554c9&builder.cachebust=true&builder.preview=symbol&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.symbol=6926ca2352d24d3286f1af71561cbb01&builder.overrides.6926ca2352d24d3286f1af71561cbb01=6926ca2352d24d3286f1af71561cbb01',
+                              hasLinks: false,
+                              kind: 'component',
+                            },
+                            priority: -98,
+                            query: [],
+                            data: {
+                              inputs: [],
+                              blocks: [
+                                {
+                                  '@type': '@builder.io/sdk:Element',
+                                  '@version': 2,
+                                  id: 'builder-4dcd66ea89804532b83d2c4f5b49cbeb',
+                                  meta: {
+                                    previousId: 'builder-dca44eb6e65b46629fad4d9beed45db3',
+                                  },
+                                  component: {
+                                    name: 'Text',
+                                    options: {
+                                      text: 'Hello world!',
+                                    },
+                                  },
+                                  responsiveStyles: {
+                                    large: {
+                                      display: 'flex',
+                                      flexDirection: 'column',
+                                      position: 'relative',
+                                      flexShrink: '0',
+                                      boxSizing: 'border-box',
+                                      marginTop: '20px',
+                                      lineHeight: 'normal',
+                                      height: 'auto',
+                                    },
+                                  },
+                                },
+                                {
+                                  '@type': '@builder.io/sdk:Element',
+                                  '@version': 2,
+                                  layerName: 'Slot',
+                                  id: 'builder-ecfff164c74a44a89f87ee03e6d2e5c4',
+                                  component: {
+                                    name: 'Slot',
+                                    options: {
+                                      name: 'children',
+                                    },
+                                  },
+                                  responsiveStyles: {
+                                    large: {
+                                      display: 'flex',
+                                      flexDirection: 'column',
+                                      position: 'relative',
+                                      flexShrink: '0',
+                                      boxSizing: 'border-box',
+                                      marginTop: '20px',
+                                    },
+                                  },
+                                },
+                                {
+                                  '@type': '@builder.io/sdk:Element',
+                                  '@version': 2,
+                                  id: 'builder-dca44eb6e65b46629fad4d9beed45db3',
+                                  meta: {
+                                    previousId: 'builder-9c9cdb77fd014a1188b16188096074cc',
+                                  },
+                                  component: {
+                                    name: 'Text',
+                                    options: {
+                                      text: 'Hello world!',
+                                    },
+                                  },
+                                  responsiveStyles: {
+                                    large: {
+                                      display: 'flex',
+                                      flexDirection: 'column',
+                                      position: 'relative',
+                                      flexShrink: '0',
+                                      boxSizing: 'border-box',
+                                      marginTop: '20px',
+                                      lineHeight: 'normal',
+                                      height: 'auto',
+                                    },
+                                  },
+                                },
+                              ],
+                            },
+                            metrics: {
+                              clicks: 0,
+                              impressions: 0,
+                            },
+                            variations: {},
+                            lastUpdated: 1705917821194,
+                            firstPublished: 1705902347406,
+                            testRatio: 1,
+                            screenshot: '',
+                            createdBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+                            lastUpdatedBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+                            folders: [],
+                          },
+                        },
+                      },
+                    },
+                  },
+                ],
+              },
+              model: 'symbol',
+              entry: '6926ca2352d24d3286f1af71561cbb01',
+              content: {
+                ownerId: 'ad30f9a246614faaa6a03374f83554c9',
+                lastUpdateBy: null,
+                createdDate: 1705902347406,
+                id: '6926ca2352d24d3286f1af71561cbb01',
+                '@version': 4,
+                name: 'hello world',
+                modelId: '5e1209efea0045f58d9b871d81b652fa',
+                published: 'published',
+                meta: {
+                  lastPreviewUrl:
+                    'https://preview.builder.codes?model=symbol&previewing=true&apiKey=ad30f9a246614faaa6a03374f83554c9&builder.space=ad30f9a246614faaa6a03374f83554c9&builder.cachebust=true&builder.preview=symbol&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.symbol=6926ca2352d24d3286f1af71561cbb01&builder.overrides.6926ca2352d24d3286f1af71561cbb01=6926ca2352d24d3286f1af71561cbb01',
+                  hasLinks: false,
+                  kind: 'component',
+                },
+                priority: -98,
+                query: [],
+                data: {
+                  inputs: [],
+                  blocks: [
+                    {
+                      '@type': '@builder.io/sdk:Element',
+                      '@version': 2,
+                      id: 'builder-4dcd66ea89804532b83d2c4f5b49cbeb',
+                      meta: {
+                        previousId: 'builder-dca44eb6e65b46629fad4d9beed45db3',
+                      },
+                      component: {
+                        name: 'Text',
+                        options: {
+                          text: 'Hello world!',
+                        },
+                      },
+                      responsiveStyles: {
+                        large: {
+                          display: 'flex',
+                          flexDirection: 'column',
+                          position: 'relative',
+                          flexShrink: '0',
+                          boxSizing: 'border-box',
+                          marginTop: '20px',
+                          lineHeight: 'normal',
+                          height: 'auto',
+                        },
+                      },
+                    },
+                    {
+                      '@type': '@builder.io/sdk:Element',
+                      '@version': 2,
+                      layerName: 'Slot',
+                      id: 'builder-ecfff164c74a44a89f87ee03e6d2e5c4',
+                      component: {
+                        name: 'Slot',
+                        options: {
+                          name: 'children',
+                        },
+                      },
+                      responsiveStyles: {
+                        large: {
+                          display: 'flex',
+                          flexDirection: 'column',
+                          position: 'relative',
+                          flexShrink: '0',
+                          boxSizing: 'border-box',
+                          marginTop: '20px',
+                        },
+                      },
+                    },
+                    {
+                      '@type': '@builder.io/sdk:Element',
+                      '@version': 2,
+                      id: 'builder-dca44eb6e65b46629fad4d9beed45db3',
+                      meta: {
+                        previousId: 'builder-9c9cdb77fd014a1188b16188096074cc',
+                      },
+                      component: {
+                        name: 'Text',
+                        options: {
+                          text: 'Hello world!',
+                        },
+                      },
+                      responsiveStyles: {
+                        large: {
+                          display: 'flex',
+                          flexDirection: 'column',
+                          position: 'relative',
+                          flexShrink: '0',
+                          boxSizing: 'border-box',
+                          marginTop: '20px',
+                          lineHeight: 'normal',
+                          height: 'auto',
+                        },
+                      },
+                    },
+                  ],
+                },
+                metrics: {
+                  clicks: 0,
+                  impressions: 0,
+                },
+                variations: {},
+                lastUpdated: 1705917821194,
+                firstPublished: 1705902347406,
+                testRatio: 1,
+                screenshot: '',
+                createdBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+                lastUpdatedBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+                folders: [],
+              },
+            },
+          },
+        },
+      },
+    ],
+  },
+  metrics: {
+    clicks: 0,
+    impressions: 0,
+  },
+  variations: {},
+  lastUpdated: 1706032649027,
+  firstPublished: 1705916526682,
+  testRatio: 1,
+  createdBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+  lastUpdatedBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+  folders: [],
+};
+
+const splitUpContent = () => {
+  const CONTENT_CLONE = JSON.parse(JSON.stringify(SLOT_WITH_SYMBOL));
+
+  const FIRST_SYMBOL_CONTENT = CONTENT_CLONE.data.blocks[0].component.options.symbol.content;
+
+  const SECOND_SYMBOL_CONTENT =
+    CONTENT_CLONE.data.blocks[0].component.options.symbol.data.children[0].component.options.symbol
+      .content;
+
+  delete CONTENT_CLONE.data.blocks[0].component.options.symbol.content;
+  delete CONTENT_CLONE.data.blocks[0].component.options.symbol.data.children[0].component.options
+    .symbol.content;
+
+  return { FIRST_SYMBOL_CONTENT, SECOND_SYMBOL_CONTENT, SLOT_WITHOUT_SYMBOL: CONTENT_CLONE };
+};
+
+export const { FIRST_SYMBOL_CONTENT, SECOND_SYMBOL_CONTENT, SLOT_WITHOUT_SYMBOL } =
+  splitUpContent();