+++ packages/sdks/src/components/content-variants/content-variants.lite.tsx
@@ -119,6 +119,7 @@ export default function ContentVariants(props: VariantsProviderProps) {
               blocksWrapperProps={props.blocksWrapperProps}
               contentWrapper={props.contentWrapper}
               contentWrapperProps={props.contentWrapperProps}
+              trustedHosts={props.trustedHosts}
             />
           )}
         </For>
@@ -146,6 +147,7 @@ export default function ContentVariants(props: VariantsProviderProps) {
         blocksWrapperProps={props.blocksWrapperProps}
         contentWrapper={props.contentWrapper}
         contentWrapperProps={props.contentWrapperProps}
+        trustedHosts={props.trustedHosts}
       />
     </>
   );